rename HscDidv2 to HscDid removed hcs-did-root-key classes and test fix
