+++ src/identity/hcs/did/hcs-did.ts
@@ -1,116 +1,190 @@
-import { PrivateKey, PublicKey, TopicId } from "@hashgraph/sdk";
-import { Hashing } from "../../../utils/hashing";
-import { DidDocumentBase } from "../../did-document-base";
+import { Client, Hbar, PrivateKey, PublicKey, Timestamp, TopicCreateTransaction, TopicId } from "@hashgraph/sdk";
+import {
+    DidDocumentBase,
+    DidMethodOperation,
+    Hashing,
+    HcsDidDidOwnerEvent,
+    HcsDidMessage,
+    HcsDidTransaction,
+    MessageEnvelope,
+} from "../../..";
 import { DidSyntax } from "../../did-syntax";
-import { HederaDid } from "../../hedera-did";
-import { HcsDidMessage } from "./hcs-did-message";
+import { HcsDidResolver } from "./hcs-did-resolver";
 
-/**
- * Hedera Decentralized Identifier for Hedera DID Method specification based on HCS.
- */
-export class HcsDid implements HederaDid {
+export class HcsDid {
     public static DID_METHOD = DidSyntax.Method.HEDERA_HCS;
 
-    private didTopicId: TopicId;
-    private network: string;
-    private idString: string;
-    private did: string;
-    private didRootKey: PublicKey;
-    private privateDidRootKey: PrivateKey;
-    private messages: HcsDidMessage[] = [];
+    protected client: Client;
+    protected privateKey: PrivateKey;
+    protected identifier: string;
+    protected network: string;
+    protected topicId: TopicId;
 
-    /**
-     * Creates a DID instance.
-     *
-     * @param network           The Hedera DID network.
-     * @param didRootKey        The public key from which DID is derived.
-     * @param didTopicId        The appnet's DID topic ID.
-     */
-    constructor(network: string, didRootKey: PublicKey, didTopicId: TopicId);
-    /**
-     * Creates a DID instance with private DID root key.
-     *
-     * @param network           The Hedera DID network.
-     * @param privateDidRootKey The private DID root key.
-     * @param didTopicId        The appnet's DID topic ID.
-     */
-    constructor(network: string, privateDidRootKey: PrivateKey, didTopicId: TopicId);
-    /**
-     * Creates a DID instance.
-     *
-     * @param network           The Hedera DID network.
-     * @param idString          The id-string of a DID.
-     * @param didTopicId        The appnet's DID topic ID.
-     * @param didRootKey        The public key from which DID is derived.
-     */
-    constructor(network: string, idString: string, didTopicId: TopicId);
-    constructor(...args: any[]) {
-        if (
-            typeof args[0] === "string" &&
-            args[1] instanceof PublicKey &&
-            (args[2] instanceof TopicId || args[2] === undefined) &&
-            args.length === 3
-        ) {
-            const [network, didRootKey, didTopicId] = args;
-            this.didTopicId = didTopicId;
-            this.network = network;
-            this.didRootKey = didRootKey;
-            this.idString = HcsDid.publicKeyToIdString(didRootKey);
-            this.did = this.buildDid();
+    protected messages: HcsDidMessage[];
+    protected resolvedAt: Timestamp;
 
-            return;
+    constructor(args: { identifier?: string; privateKey?: PrivateKey; client?: Client }) {
+        this.identifier = args.identifier;
+        this.privateKey = args.privateKey;
+        this.client = args.client;
+
+        if (!this.identifier && !this.privateKey) {
+            throw new Error("identifier and privateKey cannot be empty");
         }
 
-        if (
-            typeof args[0] === "string" &&
-            args[1] instanceof PrivateKey &&
-            (args[2] instanceof TopicId || args[2] === undefined) &&
-            args.length === 3
-        ) {
-            const [network, privateDidRootKey, didTopicId] = args;
-
-            this.didTopicId = didTopicId;
-            this.network = network;
-            this.didRootKey = privateDidRootKey.publicKey;
-            this.idString = HcsDid.publicKeyToIdString(privateDidRootKey.publicKey);
-            this.did = this.buildDid();
-            this.privateDidRootKey = privateDidRootKey;
-
-            return;
+        if (this.identifier) {
+            const [networkName, topicId] = this.parseIdentifier(this.identifier);
+            this.network = networkName;
+            this.topicId = topicId;
         }
-
-        if (
-            typeof args[0] === "string" &&
-            typeof args[1] === "string" &&
-            (args[2] instanceof TopicId || args[2] === undefined) &&
-            args.length === 3
-        ) {
-            const [network, idString, didTopicId] = args;
-
-            this.didTopicId = didTopicId;
-            this.network = network;
-            this.idString = idString;
-            this.didRootKey = HcsDid.idStringToPublicKey(idString);
-            this.did = this.buildDid();
-
-            return;
-        }
-
-        throw new Error("Couldn't find constructor");
     }
 
     /**
-     * Converts a Hedera DID string into {@link HcsDid} object.
-     *
-     * @param didString A Hedera DID string.
-     * @return {@link HcsDid} object derived from the given Hedera DID string.
+     * Public API
      */
-    public static fromString(didString: string): HcsDid {
-        if (!didString) {
-            throw new Error("DID string cannot be null");
+
+    async register() {
+        if (this.identifier) {
+            throw new Error("DID is already registered");
         }
 
-        const [didPart, topicIdPart] = didString.split(DidSyntax.DID_TOPIC_SEPARATOR);
+        if (!this.privateKey) {
+            throw new Error("privateKey is missingi i");
+        }
+
+        if (!this.client) {
+            throw new Error("Client configuration is missing");
+        }
+
+        /**
+         * Create topic
+         */
+        const topicCreateTransaction = new TopicCreateTransaction()
+            .setMaxTransactionFee(new Hbar(2))
+            .setAdminKey(this.privateKey.publicKey);
+
+        const txId = await topicCreateTransaction.execute(this.client);
+        const topicId = (await txId.getReceipt(this.client)).topicId;
+
+        this.topicId = topicId;
+        this.network = this.client.networkName;
+        this.identifier = this.buildIdentifier(this.privateKey.publicKey);
+
+        /**
+         * Set ownership
+         */
+        const event = new HcsDidDidOwnerEvent(this.identifier, this.identifier, this.privateKey.publicKey);
+        const message = new HcsDidMessage(DidMethodOperation.CREATE, this.identifier, event);
+        const envelope = new MessageEnvelope(message);
+
+        const transaction = new HcsDidTransaction(envelope, this.topicId);
+
+        // Expect some subscribtion errors because it takes time for Topic to be accessible
+
+        await new Promise((resolve, reject) => {
+            transaction
+                .signMessage((msg) => this.privateKey.sign(msg))
+                .buildAndSignTransaction((tx) => tx.setMaxTransactionFee(new Hbar(2)))
+                .onMessageConfirmed((msg) => {
+                    console.log("Submitted");
+                    resolve(msg);
+                })
+                .onError((err) => {
+                    console.log(err);
+                    reject(err);
+                })
+                .execute(this.client);
+        });
+
+        return this;
+    }
+
+    async resolve() {
+        if (!this.identifier) {
+            throw new Error("DID is not registered");
+        }
+
+        if (!this.client) {
+            throw new Error("Client configuration is missing");
+        }
+
+        return await new Promise((resolve, reject) => {
+            /**
+             * This API will have to change...
+             */
+            const resolver = new HcsDidResolver(this.topicId)
+                .setTimeout(3000)
+                .whenFinished((result) => {
+                    this.messages = result.get(this.identifier);
+
+                    let document = new DidDocumentBase(this.identifier);
+
+                    this.messages.forEach((msg) => {
+                        document = msg.getEvent().process(document);
+                    });
+
+                    resolve(document);
+                })
+                .onError((err) => {
+                    console.log(err);
+                    reject(err);
+                });
+
+            resolver.addDid(this.identifier);
+            resolver.execute(this.client);
+        });
+    }
+
+    /**
+     * Attribute getters
+     */
+
+    public getIdentifier() {
+        return this.identifier;
+    }
+
+    public getClient() {
+        return this.client;
+    }
+
+    public getPrivateKey() {
+        return this.privateKey;
+    }
+
+    public getTopicId() {
+        return this.topicId;
+    }
+
+    public getNetwork() {
+        return this.network;
+    }
+
+    public getMethod() {
+        return HcsDid.DID_METHOD;
+    }
+
+    /**
+     * Private
+     */
+
+    private buildIdentifier(publicKey: PublicKey): string {
+        const methodNetwork = [this.getMethod().toString(), this.network].join(DidSyntax.DID_METHOD_SEPARATOR);
+
+        let ret: string;
+        ret =
+            DidSyntax.DID_PREFIX +
+            DidSyntax.DID_METHOD_SEPARATOR +
+            methodNetwork +
+            DidSyntax.DID_METHOD_SEPARATOR +
+            this.publicKeyToIdString(publicKey) +
+            DidSyntax.DID_TOPIC_SEPARATOR +
+            this.topicId.toString();
+
+        return ret;
+    }
+
+    private parseIdentifier(identifier: string): [string, TopicId] {
+        const [didPart, topicIdPart] = identifier.split(DidSyntax.DID_TOPIC_SEPARATOR);
 
         if (!topicIdPart) {
             throw new Error("DID string is invalid: topic ID is missing");
@@ -142,123 +216,13 @@ export class HcsDid implements HederaDid {
                 throw new Error("DID string is invalid.");
             }
 
-            return new HcsDid(networkName, didIdString, topicId);
+            return [networkName, topicId];
         } catch (e) {
             throw new Error("DID string is invalid. " + e.message);
         }
     }
 
-    /**
-     * Generates a random DID root key.
-     *
-     * @return A private key of generated public DID root key.
-     */
-    public static generateDidRootKey(): PrivateKey {
-        return PrivateKey.generate();
-    }
-
-    /**
-     * Generates DID document base from the given DID and its root key.
-     *
-     * @param didRootKey Public key used to build this DID.
-     * @return The DID document base.
-     * @throws IllegalArgumentException In case given DID root key does not match this DID.
-     */
-    public generateDidDocument(): DidDocumentBase {
-        let result = new DidDocumentBase(this.toDid());
-
-        // proccess events to get uptodate did document
-        this.getMessages().forEach((msg) => {
-            result = msg.getEvent().process(result);
-        });
-
-        return result;
-    }
-
-    public getNetwork(): string {
-        return this.network;
-    }
-
-    public getMethod(): DidSyntax.Method {
-        return DidSyntax.Method.HEDERA_HCS;
-    }
-
-    public toString(): string {
-        return this.did;
-    }
-
-    public getDidTopicId(): TopicId {
-        return this.didTopicId;
-    }
-
-    public getIdString(): string {
-        return this.idString;
-    }
-
-    public toDid() {
-        return this.did;
-    }
-
-    public getMessages() {
-        return this.messages;
-    }
-
-    public addMessage(message: HcsDidMessage) {
-        this.messages.push(message);
-    }
-
-    /**
-     * Constructs DID string from the instance of DID object.
-     *
-     * @return A DID string.
-     */
-    private buildDid(): string {
-        const methodNetwork = [this.getMethod().toString(), this.network].join(DidSyntax.DID_METHOD_SEPARATOR);
-
-        let ret: string;
-        ret =
-            DidSyntax.DID_PREFIX +
-            DidSyntax.DID_METHOD_SEPARATOR +
-            methodNetwork +
-            DidSyntax.DID_METHOD_SEPARATOR +
-            this.idString +
-            DidSyntax.DID_TOPIC_SEPARATOR +
-            this.didTopicId.toString();
-
-        return ret;
-    }
-
-    /**
-     * Constructs an id-string of a DID from a given public key.
-     *
-     * @param didRootKey Public Key from which the DID is created.
-     * @return The id-string of a DID that is a Base58-encoded SHA-256 hash of a given public key.
-     */
-    public static publicKeyToIdString(didRootKey: PublicKey): string {
-        return Hashing.multibase.encode(didRootKey.toBytes());
-    }
-
-    public static idStringToPublicKey(idString: string): PublicKey {
-        return PublicKey.fromBytes(Hashing.multibase.decode(idString));
-    }
-
-    /**
-     * Returns a private key of DID root key.
-     * This is only available if it was provided during {@link HcsDid} construction.
-     *
-     * @return The private key of DID root key.
-     */
-    public getPrivateDidRootKey(): PrivateKey {
-        return this.privateDidRootKey;
-    }
-
-    /**
-     * Returns a public key of DID root key.
-     * This is only available if it was provided during {@link HcsDid} construction.
-     *
-     * @return The private key of DID root key.
-     */
-    public getPublicDidRootKey(): PublicKey {
-        return this.didRootKey;
+    private publicKeyToIdString(publicKey: PublicKey): string {
+        return Hashing.multibase.encode(publicKey.toBytes());
     }
 }
