+++ src/index.ts
@@ -10,10 +10,8 @@ import { HcsDidVerificationRelationshipEvent } from "./identity/hcs/did/event/hc
 import { HcsDid } from "./identity/hcs/did/hcs-did";
 import { HcsDidMessage } from "./identity/hcs/did/hcs-did-message";
 import { HcsDidResolver } from "./identity/hcs/did/hcs-did-resolver";
-import { HcsDidRootKey } from "./identity/hcs/did/hcs-did-root-key";
 import { HcsDidTopicListener } from "./identity/hcs/did/hcs-did-topic-listener";
 import { HcsDidTransaction } from "./identity/hcs/did/hcs-did-transaction";
-import { HcsDidV2 } from "./identity/hcs/did/hcs-did-v2";
 import { JsonClass } from "./identity/hcs/json-class";
 import { Message } from "./identity/hcs/message";
 import { MessageEnvelope } from "./identity/hcs/message-envelope";
@@ -38,10 +36,8 @@ export {
     DidSyntax,
     Hashing,
     HcsDid,
-    HcsDidV2,
     HcsDidMessage,
     HcsDidResolver,
-    HcsDidRootKey,
     HcsDidTopicListener,
     HcsDidTransaction,
     HcsDidDidOwnerEvent,
