+++ test/did/hcs-did.js
@@ -1,84 +1,202 @@
-const { TopicId } = require("@hashgraph/sdk");
-const { DidSyntax, HcsDid, Hashing } = require("../../dist");
+const { AccountId, PrivateKey, Client, Timestamp, TopicMessageQuery } = require("@hashgraph/sdk");
+const { OPERATOR_KEY, OPERATOR_ID, NETWORK } = require("../variables");
+const { HcsDid, Hashing } = require("../../dist");
 
 const { assert } = require("chai");
 
-const network = "testnet";
+const TOPIC_REGEXP = /^0\.0\.[0-9]{8,}/;
 
 describe("HcsDid", function () {
-    it("Test Generate And Parse Did With Tid", async function () {
-        const didTopicId = "1.5.23462345";
+    describe("#constructor", () => {
+        it("throws error because of missing identifier and privateKey", () => {
+            assert.throw(() => {
+                new HcsDid({});
+            });
+        });
 
-        const privateKey = HcsDid.generateDidRootKey();
+        it("successfuly builds HcsDid with private key only", () => {
+            const privateKey = PrivateKey.generate();
+            const did = new HcsDid({ privateKey });
 
-        const topicId = TopicId.fromString(didTopicId);
-        const did = new HcsDid(network, privateKey.publicKey, topicId);
+            assert.equal(did.getIdentifier(), null);
+            assert.equal(did.getPrivateKey(), privateKey);
+            assert.equal(did.getClient(), null);
+            assert.equal(did.getTopicId(), null);
+            assert.equal(did.getNetwork(), null);
+        });
 
-        const didString = did.toString();
+        it("successfuly builds HcsDid with identifier only", () => {
+            const identifier = "did:hedera:testnet:z6MkgUv5CvjRP6AsvEYqSRN7djB6p4zK9bcMQ93g5yK6Td7N_0.0.29613327";
+            const did = new HcsDid({ identifier });
 
-        assert.exists(didString);
+            assert.equal(did.getIdentifier(), identifier);
+            assert.equal(did.getPrivateKey(), null);
+            assert.equal(did.getClient(), null);
+            assert.equal(did.getTopicId(), "0.0.29613327");
+            assert.equal(did.getNetwork(), "testnet");
+        });
 
-        const parsedDid = HcsDid.fromString(didString);
+        it("throws error if passed identifer is invalid", () => {
+            [
+                null,
+                "invalidDid1",
+                "did:invalid",
+                "did:invalidMethod:8LjUL78kFVnWV9rFnNCTE5bZdRmjm2obqJwS892jVLak_0.0.24352",
+                "did:hedera:invalidNetwork:8LjUL78kFVnWV9rFnNCTE5bZdRmjm2obqJwS892jVLak_0.0.24352",
+                "did:hedera:testnet:invalidAddress_0.0.24352_1.5.23462345",
+                "did:hedera:testnet_1.5.23462345",
+                "did:hedera:testnet:z6Mk8LjUL78kFVnWV9rFnNCTE5bZdRmjm2obqJwS892jVLak:unknown:parameter=1_missing",
+                "did:hedera:testnet:z6Mk8LjUL78kFVnWV9rFnNCTE5bZdRmjm2obqJwS892jVLak_0.0.1=1",
+                "did:hedera:testnet:z6Mk8LjUL78kFVnWV9rFnNCTE5bZdRmjm2obqJwS892jVLak:hedera:testnet:fid",
+                "did:hedera:testnet:z6Mk8LjUL78kFVnWV9rFnNCTE5bZdRmjm2obqJwS892jVLak:unknownPart_0.0.1",
+                "did:notHedera:testnet:z6Mk8LjUL78kFVnWV9rFnNCTE5bZdRmjm2obqJwS892jVLak_0.0.1",
+            ].forEach((identifier) => {
+                assert.throw(() => {
+                    new HcsDid({ identifier });
+                });
+            });
+        });
 
-        assert.exists(parsedDid);
-        assert.exists(parsedDid.getDidTopicId());
+        it("accepts client parameter", () => {
+            const client = Client.forTestnet();
+            const identifier = "did:hedera:testnet:z6MkgUv5CvjRP6AsvEYqSRN7djB6p4zK9bcMQ93g5yK6Td7N_0.0.29613327";
+            const did = new HcsDid({ identifier, client });
 
-        assert.equal(parsedDid.toDid(), didString);
-        assert.equal(parsedDid.getMethod(), DidSyntax.Method.HEDERA_HCS);
-        assert.equal(parsedDid.getNetwork(), network);
-        assert.equal(parsedDid.getDidTopicId().toString(), didTopicId);
-        assert.equal(parsedDid.getIdString(), did.getIdString());
-
-        const parsedDocument = parsedDid.generateDidDocument();
-
-        assert.exists(parsedDocument);
-        assert.equal(parsedDocument.getId(), parsedDid.toString());
-        assert.equal(parsedDocument.getContext(), DidSyntax.DID_DOCUMENT_CONTEXT);
-        assert.notExists(parsedDocument.getDidRootKey());
-
-        const document = did.generateDidDocument();
-
-        assert.exists(document);
-        assert.equal(document.getId(), parsedDid.toString());
-        assert.equal(document.getContext(), DidSyntax.DID_DOCUMENT_CONTEXT);
-        assert.exists(document.getDidRootKey());
-        assert.equal(
-            document.getDidRootKey().getPublicKeyMultibase(),
-            Hashing.multibase.encode(privateKey.publicKey.toBytes())
-        );
+            assert.equal(did.getIdentifier(), identifier);
+            assert.equal(did.getPrivateKey(), null);
+            assert.equal(did.getClient(), client);
+            assert.equal(did.getTopicId(), "0.0.29613327");
+            assert.equal(did.getNetwork(), "testnet");
+        });
     });
 
-    it("Test Parse Predefined Dids", async function () {
-        const didTopicId = "1.5.23462345";
+    describe("#register", async () => {
+        let client;
 
-        const validDidWithSwitchedParamsOrder =
-            "did:hedera:testnet:z6Mk8LjUL78kFVnWV9rFnNCTE5bZdRmjm2obqJwS892jVLak" + "_" + didTopicId;
+        before(async () => {
+            const operatorId = AccountId.fromString(OPERATOR_ID);
+            const operatorKey = PrivateKey.fromString(OPERATOR_KEY);
+            client = Client.forTestnet();
+            client.setMirrorNetwork(["hcs." + NETWORK + ".mirrornode.hedera.com:5600"]);
+            client.setOperator(operatorId, operatorKey);
+        });
 
-        const invalidDids = [
-            null,
-            "invalidDid1",
-            "did:invalid",
-            "did:invalidMethod:8LjUL78kFVnWV9rFnNCTE5bZdRmjm2obqJwS892jVLak_0.0.24352",
-            "did:hedera:invalidNetwork:8LjUL78kFVnWV9rFnNCTE5bZdRmjm2obqJwS892jVLak_0.0.24352",
-            "did:hedera:testnet:invalidAddress_0.0.24352_1.5.23462345",
-            "did:hedera:testnet_1.5.23462345",
-            "did:hedera:testnet:z6Mk8LjUL78kFVnWV9rFnNCTE5bZdRmjm2obqJwS892jVLak:unknown:parameter=1_missing",
-            "did:hedera:testnet:z6Mk8LjUL78kFVnWV9rFnNCTE5bZdRmjm2obqJwS892jVLak_0.0.1=1",
-            "did:hedera:testnet:z6Mk8LjUL78kFVnWV9rFnNCTE5bZdRmjm2obqJwS892jVLak:hedera:testnet:fid",
-            "did:hedera:testnet:z6Mk8LjUL78kFVnWV9rFnNCTE5bZdRmjm2obqJwS892jVLak:unknownPart_0.0.1",
-            "did:notHedera:testnet:z6Mk8LjUL78kFVnWV9rFnNCTE5bZdRmjm2obqJwS892jVLak_0.0.1",
-        ];
+        it("throws error if DID is already registered", async () => {
+            const identifier = "did:hedera:testnet:z6MkgUv5CvjRP6AsvEYqSRN7djB6p4zK9bcMQ93g5yK6Td7N_0.0.29613327";
+            const did = new HcsDid({ identifier });
 
-        for (let did of invalidDids) {
-            assert.throw(() => {
-                HcsDid.fromString(did);
+            try {
+                await did.register();
+            } catch (err) {
+                assert.instanceOf(err, Error);
+                assert.equal(err.message, "DID is already registered");
+            }
+        });
+
+        it("throws error if client configuration is missing", async () => {
+            const privateKey = PrivateKey.generate();
+            const did = new HcsDid({ privateKey });
+
+            try {
+                await did.register();
+            } catch (err) {
+                assert.instanceOf(err, Error);
+                assert.equal(err.message, "Client configuration is missing");
+            }
+        });
+
+        it("creates new DID by registering a topic and submitting first message", async () => {
+            const privateKey = PrivateKey.fromString(OPERATOR_KEY);
+            const did = new HcsDid({ privateKey, client });
+            const result = await did.register();
+
+            assert.equal(result, did);
+            assert.match(did.getTopicId().toString(), TOPIC_REGEXP);
+            assert.equal(
+                did.getIdentifier(),
+                `did:hedera:testnet:${Hashing.multibase.encode(privateKey.publicKey.toBytes())}_${did
+                    .getTopicId()
+                    .toString()}`
+            );
+            assert.equal(did.getPrivateKey(), privateKey);
+            assert.equal(did.getClient(), client);
+            assert.equal(did.getNetwork(), "testnet");
+
+            const messages = [];
+
+            new TopicMessageQuery()
+                .setTopicId(did.getTopicId())
+                .setStartTime(new Timestamp(0, 0))
+                .setEndTime(Timestamp.fromDate(new Date()))
+                .subscribe(client, (msg) => {
+                    messages.push(msg);
+                });
+
+            /**
+             * Will have to change, let's just wait for 3s and assume all messages were read
+             */
+            await new Promise((resolve) => setTimeout(resolve, 3000));
+
+            assert.equal(messages.length, 1);
+        }).timeout(60000);
+    });
+
+    describe("#resolve", () => {
+        let client;
+
+        before(async () => {
+            const operatorId = AccountId.fromString(OPERATOR_ID);
+            const operatorKey = PrivateKey.fromString(OPERATOR_KEY);
+            client = Client.forTestnet();
+            client.setMirrorNetwork(["hcs." + NETWORK + ".mirrornode.hedera.com:5600"]);
+            client.setOperator(operatorId, operatorKey);
+        });
+
+        it("throws error about unregistered DID", async () => {
+            const privateKey = PrivateKey.generate();
+            const did = new HcsDid({ privateKey, client });
+
+            try {
+                await did.resolve();
+            } catch (err) {
+                assert.instanceOf(err, Error);
+                assert.equal(err.message, "DID is not registered");
+            }
+        });
+
+        it("throws error about missing Client parameter", async () => {
+            const identifier = "did:hedera:testnet:z6MkgUv5CvjRP6AsvEYqSRN7djB6p4zK9bcMQ93g5yK6Td7N_0.0.29613327";
+            const did = new HcsDid({ identifier });
+
+            try {
+                await did.resolve();
+            } catch (err) {
+                assert.instanceOf(err, Error);
+                assert.equal(err.message, "Client configuration is missing");
+            }
+        });
+
+        it("successfuly resolves just registered DID", async () => {
+            const privateKey = PrivateKey.fromString(OPERATOR_KEY);
+            const did = new HcsDid({ privateKey, client });
+
+            await did.register();
+
+            const didDocument = (await did.resolve()).toJsonTree();
+
+            assert.equal(didDocument["@context"], "https://www.w3.org/ns/did/v1");
+            assert.equal(didDocument["id"], did.getIdentifier());
+            assert.equal(didDocument["assertionMethod"].length, 1);
+            assert.equal(didDocument["assertionMethod"][0], `${did.getIdentifier()}#did-root-key`);
+            assert.equal(didDocument["authentication"].length, 1);
+            assert.equal(didDocument["authentication"][0], `${did.getIdentifier()}#did-root-key`);
+            assert.equal(didDocument["verificationMethod"].length, 1);
+            assert.deepEqual(didDocument["verificationMethod"][0], {
+                id: `${did.getIdentifier()}#did-root-key`,
+                type: "Ed25519VerificationKey2018",
+                controller: did.getIdentifier(),
+                publicKeyMultibase: Hashing.multibase.encode(privateKey.publicKey.toBytes()),
             });
-        }
-
-        const validDid = HcsDid.fromString(validDidWithSwitchedParamsOrder);
-
-        assert.exists(validDid);
-        assert.exists(validDid.getDidTopicId());
-        assert.equal(validDid.getDidTopicId().toString(), didTopicId);
+        }).timeout(60000);
     });
 });
